PYIC-3591: Use local config specifically for the reuse page
Repurposed the existing unused data we already had defined
Refactored to move the reuse code into the helper as it is now used in two places
